backend/requirements.txt
# all the dependency are the default of pip, python:3.12-bookworm, 26/11/2024

# Django and (from Django) dependency
Django==5.1.3
asgiref==3.8.1
sqlparse==0.5.2

# djangorestframework
djangorestframework==3.15.2
django-cors-headers

# optional dependency from https://www.django-rest-framework.org/#quickstart
PyYAML==6.0.2 #uritemplate==4.1.1
Markdown==3.7
Pygments==2.18.0
django-filter==24.3
django-guardian==2.4.0

#for db.backends.postgresql
psycopg2==2.9.10

#for JWT (https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html#installation)
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1

# for TOTP/2FA
pyotp==2.9.0
qrcode[pil]==8.0

# for OAuth
requests==2.31.0

# for blockchain
web3==6.10.0
django-environ==0.11.2

# for websockets
channels==4.2.0

# for daphne
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
service-identity==24.2.0
setuptools==75.6.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.2
